refactor(core): reorg the sso utils
reorg the sso utils
