+++ packages/core/src/routes/sso-connector/utils.test.ts
@@ -12,19 +12,7 @@ await mockEsmWithActual('#src/sso/OidcConnector/utils.js', () => ({
 }));
 
 const { ssoConnectorFactories } = await import('#src/sso/index.js');
-const { isSupportedSsoProvider, parseFactoryDetail, fetchConnectorProviderDetails } = await import(
-  './utils.js'
-);
-
-describe('isSupportedSsoProvider', () => {
-  it.each(Object.values(SsoProviderName))('should return true for %s', (providerName) => {
-    expect(isSupportedSsoProvider(providerName)).toBe(true);
-  });
-
-  it('should return false for unknown provider', () => {
-    expect(isSupportedSsoProvider('unknown-provider')).toBe(false);
-  });
-});
+const { parseFactoryDetail, fetchConnectorProviderDetails } = await import('./utils.js');
 
 describe('parseFactoryDetail', () => {
   it.each(Object.values(SsoProviderName))('should return correct detail for %s', (providerName) => {
