+++ packages/core/src/sso/utils.test.ts
@@ -1,0 +1,12 @@
+import { SsoProviderName } from './types/index.js';
+import { isSupportedSsoProvider } from './utils.js';
+
+describe('isSupportedSsoProvider', () => {
+  it.each(Object.values(SsoProviderName))('should return true for %s', (providerName) => {
+    expect(isSupportedSsoProvider(providerName)).toBe(true);
+  });
+
+  it('should return false for unknown provider', () => {
+    expect(isSupportedSsoProvider('unknown-provider')).toBe(false);
+  });
+});
